Extract node traverse to new file
